requirements.txt
amqp==5.0.5
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
lxml==4.6.2
MarkupSafe==1.1.1
packaging==20.9
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
querystring-parser==1.2.4
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
